% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichHarmonizome.R
\name{EnrichHarmonizome}
\alias{EnrichHarmonizome}
\title{EnrichHarmonizome}
\usage{
EnrichHarmonizome(
  gene,
  tbl,
  pvalueCutoff,
  pAdjustMethod,
  minGSSize = 10,
  maxGSSize = 100,
  qvalueCutoff
)
}
\arguments{
\item{gene}{A character vector containing the gene names.}

\item{tbl}{A data frame containing the gene sets in the columns and gene names in the rows.}

\item{pvalueCutoff}{A numeric value for the p-value cutoff used for the enrichment analysis.}

\item{pAdjustMethod}{A character string indicating the method used for adjusting p-values.}

\item{minGSSize}{An integer specifying the minimum number of genes in a gene set for it to be considered for the enrichment analysis.}

\item{maxGSSize}{An integer specifying the maximum number of genes in a gene set for it to be considered for the enrichment analysis.}

\item{qvalueCutoff}{A numeric value for the q-value cutoff used for the enrichment analysis.}
}
\value{
Returns a gene set enrichment analysis result object generated by the enricher function from the clusterProfiler package.
}
\description{
Given a set of genes, a gene set database, and enrichment parameters, performs gene set enrichment analysis using the clusterProfiler package.
}
\examples{
EnrichHarmonizome(gene = c("PTEN", "TP53"), tbl = example_tbl, pvalueCutoff = 0.05, pAdjustMethod = "BH", qvalueCutoff = 0.05)

}
\seealso{
\code{\link{enricher}} function from the \code{clusterProfiler} package.
}
