% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_to_list.R
\name{geneset_to_list}
\alias{geneset_to_list}
\title{Gene Set to List}
\usage{
geneset_to_list(tbl)
}
\arguments{
\item{tbl}{A tibble containing gene IDs or other gene-related data.}
}
\value{
A named list of vectors containing the gene IDs or other gene-related data.
}
\description{
This function takes a tibble and converts it into a named list of vectors.
The tibble should contain a "name" column and any number of additional columns containing gene IDs or other gene-related data.
The resulting list will have names equal to the values in the "name" column and will contain vectors of the remaining columns' values.
This function makes use of the tibble, dplyr, tidyr, and purrr packages.
}
\examples{
data("my_geneset")
geneset_to_list(my_geneset)

}
\seealso{
https://tibble.tidyverse.org/

https://dplyr.tidyverse.org/

https://tidyr.tidyverse.org/

https://purrr.tidyverse.org/
}
\keyword{gene,}
\keyword{list}
\keyword{tibble,}
