% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneset.R
\name{get_geneset}
\alias{get_geneset}
\title{Get Gene Set}
\usage{
get_geneset(code)
}
\arguments{
\item{code}{A character vector or a string specifying the Harmonizome dataset code(s) for which to retrieve the gene sets.}
}
\value{
A tibble with columns for dataset code and gene IDs in each gene set.
}
\description{
This function retrieves the gene sets contained in one or more Harmonizome datasets based on the given code(s).
The function requires the dplyr, httr, and jsonlite packages.
It first filters the datasets based on the given codes and then retrieves the gene sets for each dataset using the Harmonizome API.
The resulting tibble contains information about the gene sets, including the dataset code and the gene IDs in each gene set.
}
