% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAHarmonizome.R
\name{GSEAHarmonizome}
\alias{GSEAHarmonizome}
\title{Run Gene Set Enrichment Analysis (GSEA) using Harmonizome gene sets}
\usage{
GSEAHarmonizome(pathways, stats, minSize, maxSize)
}
\arguments{
\item{pathways}{A named list of gene sets, where each element is a vector of
gene symbols representing a gene set. The names of the list
should correspond to the names of the gene sets.}

\item{stats}{A vector of numeric gene expression values.}

\item{minSize}{The minimum size of the gene set to be considered in the analysis.
Default is 1.}

\item{maxSize}{The maximum size of the gene set to be considered in the analysis.
Default is 5000.}
}
\value{
A data frame containing the GSEA results, including the gene set name,
the size of the gene set, the enrichment score, and the p-value.
}
\description{
This function performs Gene Set Enrichment Analysis (GSEA) using gene sets
from Harmonizome. It uses the fgsea package to calculate enrichment scores
and p-values for the given gene set and gene expression data.
}
\examples{
library(dplyr)
data(cancer_eset)
gene_symbols <- as.character(rownames(exprs(cancer_eset)))
gseah_genesets <- get_geneset(c("GO_Biological_Process_2018", "KEGG_2019_Human"))
gseah_geneset_list <- geneset_to_list(gseah_genesets)
gseah_results <- GSEAHarmonizome(gseah_geneset_list, gene_symbols, minSize = 10, maxSize = 100)

}
