% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSVAHarmonizome.R
\name{GSVAHarmonizome}
\alias{GSVAHarmonizome}
\title{Run GSVA analysis using Harmonizome pathways}
\usage{
GSVAHarmonizome(expr, pathways)
}
\arguments{
\item{expr}{A numeric matrix or data frame with gene expression values.}

\item{pathways}{A list of character vectors, each representing a gene set in the Harmonizome.}
}
\value{
A numeric matrix with GSVA enrichment scores for each pathway.
}
\description{
Run GSVA analysis using Harmonizome pathways
}
\examples{
data(example_expr)

result <- GSVAHarmonizome(expr = example_expr, pathways = example_pathways)
}
\keyword{GSVA,}
\keyword{analysis}
\keyword{pathway}
