% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{harmonizome_datasets}
\alias{harmonizome_datasets}
\title{Harmonizome dataset collection}
\format{
## `harmonizome_datasets`
A data frame with 129 rows and 3 columns:
\describe{
  \item{Name}{Dataset name}
  \item{Code}{Dataset codes}
  \item{Link}{Dataset Link}
  ...
}
}
\source{
<https://maayanlab.cloud/Harmonizome/api/1.0/dataset/>
}
\usage{
harmonizome_datasets
}
\description{
The full collection available of Harmonizome API
}
\keyword{datasets}
