% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_dataset_collection.R
\name{show_dataset_collection}
\alias{show_dataset_collection}
\title{Show Dataset Collection}
\usage{
show_dataset_collection()
}
\value{
A table with information about the Harmonizome dataset collection.
}
\description{
This function loads and displays information about the datasets available in the Harmonizome collection.
The information is displayed in a table format, and the function also provides a link to the complete dataset information available online.
}
\examples{
show_dataset_collection()

}
\seealso{
https://maayanlab.cloud/Harmonizome/
}
\keyword{dataset,}
\keyword{harmonizome}
