% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEAHarmonizome.R
\name{ssGSEAHarmonizome}
\alias{ssGSEAHarmonizome}
\title{Run single-sample GSEA (ssGSEA) analysis using Harmonizome pathways}
\usage{
ssGSEAHarmonizome(expr, pathways)
}
\arguments{
\item{expr}{A numeric matrix or data frame with gene expression values.}

\item{pathways}{A list of character vectors, each representing a gene set in the Harmonizome.}
}
\value{
A numeric matrix with ssGSEA enrichment scores for each pathway in each sample.
}
\description{
This function performs single-sample GSEA (ssGSEA) analysis using gene expression data
and gene sets from Harmonizome. It utilizes the GSVA package to calculate enrichment scores
for each gene set in each sample.
}
\examples{
data(example_expr)
data(example_pathways)
result <- ssGSEA(expr = example_expr, pathways = example_pathways)

}
